Refactor: Rename `Playwright.Vue` to `Playwright.WebApp.Vue` across files and configurations for improved clarity and consistency.
