+++ .gitignore
@@ -352,7 +352,7 @@ healthchecksdb
 # wwwroot folders from js web based projects
 /examples/InfiniFrameExample.WebApp.React/wwwroot/
 /examples/InfiniFrameExample.WebApp.Vue/wwwroot/
-/tests/InfiniFrameTests.Playwright.Vue/wwwroot/
+/tests/InfiniFrameTests.Playwright.WebApp.Vue/wwwroot/
 #/tests/InfiniFrameTests.Playwright.BlazorWebView/wwwroot/ This needs to be included for Blazor to work
 
 /docs/node_modules/
