+++ InfiniFrame.GitHubActions.slnf
@@ -1,4 +1,4 @@
-﻿{
+{
     "solution": {
         "path": "InfiniFrame.slnx",
         "projects": [
@@ -11,7 +11,7 @@
             "src\\InfiniFrame.WebServer\\InfiniFrame.WebServer.csproj",
             "src\\InfiniFrame.Native\\InfiniFrame.Native.proj",
             "tests\\InfiniFrameTests\\InfiniFrameTests.csproj",
-            "tests\\InfiniFrameTests.Playwright.Vue\\InfiniFrameTests.Playwright.Vue.csproj",
+            "tests\\InfiniFrameTests.Playwright.WebApp.Vue\\InfiniFrameTests.Playwright.WebApp.Vue.csproj",
             "tests\\InfiniFrameTests.Playwright.BlazorWebView\\InfiniFrameTests.Playwright.BlazorWebView.csproj",
             "tests\\InfiniFrameTests.Shared\\InfiniFrameTests.Shared.csproj",
             "tests\\InfiniFrameTests.Tools.Pack\\InfiniFrameTests.Tools.Pack.csproj",
