+++ tests/InfiniFrameTests.Playwright.WebApp.Vue/JavascriptInteropTests.cs
@@ -1,13 +1,17 @@
-﻿// ---------------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
-using InfiniFrameTests.Playwright.Vue.TestUtility;
+using InfiniFrameTests.Playwright.WebApp.Vue.TestUtility;
 
-namespace InfiniFrameTests.Playwright.Vue;
+namespace InfiniFrameTests.Playwright.WebApp.Vue;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
 [InheritsTests]
-public sealed class WebviewWindowTests : SharedWebviewWindowTests {
+public sealed class JavascriptInteropTests : SharedJavascriptInteropTests {
     protected override IPlaywrightRuntimeContext RuntimeContext => VuePlaywrightRuntimeContext.Instance;
+
+    protected override string FullscreenToggleButtonSelector => "#fullscreen-toggle-button";
+
+    protected override string TitleToggleButtonSelector => "#title-toggle-button";
 }
