+++ tests/InfiniFrameTests.Playwright.WebApp.Vue/Sources/infiniframe-playwright-vue/src/components/TitleChangeButton.vue
