+++ tests/InfiniFrameTests.Playwright.WebApp.Vue/Sources/infiniframe-playwright-vue/vite.config.ts
