+++ tests/InfiniFrameTests.Playwright.WebApp.Vue/TestUtility/GlobalPlaywrightContext.cs
@@ -1,4 +1,4 @@
-﻿// ---------------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
 using InfiniFrame;
@@ -8,7 +8,7 @@ using System.Net;
 using System.Net.Sockets;
 using InfiniFrame.Js.Interop.MessageHandlers;
 
-namespace InfiniFrameTests.Playwright.Vue.TestUtility;
+namespace InfiniFrameTests.Playwright.WebApp.Vue.TestUtility;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
